python-services/video/requirements.txt
fastapi
uvicorn
requests
python-dotenv
ffmpeg-python